add -W all to makefiles

+++ arm9/src/filesystem.c
@@ -172,27 +172,27 @@ THUMB_FUNC void ReadFromNarcMemberByPathAndId(void * dest, const char * path, s3
     FS_SeekFile(&file, 12, FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 2);
     chunk_starts[0] = chunk_size;
-    FS_SeekFile(&file, chunk_starts[0] + 4, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[0] + 4), FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 4);
     FS_ReadFile(&file, &num_files, 2);
     if (num_files <= file_idx)
         ErrorHandling();
     chunk_starts[1] = chunk_starts[0] + chunk_size;
-    FS_SeekFile(&file, chunk_starts[1] + 4, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[1] + 4), FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 4);
     btnf_size = chunk_size;
-    FS_SeekFile(&file, chunk_starts[0] + 12 + 8 * file_idx, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[0] + 12 + 8 * file_idx), FS_SEEK_SET);
     FS_ReadFile(&file, &file_start, 4);
     FS_ReadFile(&file, &file_end, 4);
     chunk_starts[2] = chunk_starts[1] + btnf_size;
-    FS_SeekFile(&file, chunk_starts[2] + 8 + file_start + offset, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[2] + 8 + file_start + offset), FS_SEEK_SET);
     if (size == 0)
         chunk_size = file_end - file_start;
     else
         chunk_size = size;
     if (chunk_size == 0)
         ErrorHandling();
-    FS_ReadFile(&file, dest, chunk_size);
+    FS_ReadFile(&file, dest, (s32)chunk_size);
     FS_CloseFile(&file);
 }
 
@@ -213,20 +213,20 @@ THUMB_FUNC void * AllocAndReadFromNarcMemberByPathAndId(const char * path, s32 f
     FS_SeekFile(&file, 12, FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 2);
     chunk_starts[0] = chunk_size;
-    FS_SeekFile(&file, chunk_starts[0] + 4, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[0] + 4), FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 4);
     FS_ReadFile(&file, &num_files, 2);
     if (num_files <= file_idx)
         ErrorHandling();
     chunk_starts[1] = chunk_starts[0] + chunk_size;
-    FS_SeekFile(&file, chunk_starts[1] + 4, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[1] + 4), FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 4);
     btnf_size = chunk_size;
-    FS_SeekFile(&file, chunk_starts[0] + 12 + 8 * file_idx, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[0] + 12 + 8 * file_idx), FS_SEEK_SET);
     FS_ReadFile(&file, &file_start, 4);
     FS_ReadFile(&file, &file_end, 4);
     chunk_starts[2] = chunk_starts[1] + btnf_size;
-    FS_SeekFile(&file, chunk_starts[2] + 8 + file_start + offset, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[2] + 8 + file_start + offset), FS_SEEK_SET);
     if (size == 0)
         chunk_size = file_end - file_start;
     else
@@ -242,7 +242,7 @@ THUMB_FUNC void * AllocAndReadFromNarcMemberByPathAndId(const char * path, s32 f
         dest = AllocFromHeapAtEnd(heap_id, chunk_size);
         break;
     }
-    FS_ReadFile(&file, dest, chunk_size);
+    FS_ReadFile(&file, dest, (s32)chunk_size);
     FS_CloseFile(&file);
     return dest;
 }
@@ -293,20 +293,20 @@ THUMB_FUNC u32 GetNarcMemberSizeByIdPair(NarcId narc_id, s32 file_idx)
     FS_SeekFile(&file, 12, FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 2);
     chunk_starts[0] = chunk_size;
-    FS_SeekFile(&file, chunk_starts[0] + 4, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[0] + 4), FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 4);
     FS_ReadFile(&file, &num_files, 2);
     if (num_files <= file_idx)
         ErrorHandling();
     chunk_starts[1] = chunk_starts[0] + chunk_size;
-    FS_SeekFile(&file, chunk_starts[1] + 4, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[1] + 4), FS_SEEK_SET);
     FS_ReadFile(&file, &chunk_size, 4);
     btnf_size = chunk_size;
-    FS_SeekFile(&file, chunk_starts[0] + 12 + 8 * file_idx, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[0] + 12 + 8 * file_idx), FS_SEEK_SET);
     FS_ReadFile(&file, &file_start, 4);
     FS_ReadFile(&file, &file_end, 4);
     chunk_starts[2] = chunk_starts[1] + btnf_size;
-    FS_SeekFile(&file, chunk_starts[2] + 8 + file_start + 0, FS_SEEK_SET);
+    FS_SeekFile(&file, (s32)(chunk_starts[2] + 8 + file_start + 0), FS_SEEK_SET);
     chunk_size = file_end - file_start;
     if (chunk_size == 0)
         ErrorHandling();
@@ -326,11 +326,11 @@ THUMB_FUNC NARC * NARC_ctor(NarcId narc_id, u32 heap_id)
         FS_OpenFile(&narc->file, sNarcFileList[narc_id]);
         FS_SeekFile(&narc->file, 12, FS_SEEK_SET);
         FS_ReadFile(&narc->file, &narc->btaf_start, 2);
-        FS_SeekFile(&narc->file, narc->btaf_start + 4, FS_SEEK_SET);
+        FS_SeekFile(&narc->file, (s32)(narc->btaf_start + 4), FS_SEEK_SET);
         FS_ReadFile(&narc->file, &chunk_size, 4);
         FS_ReadFile(&narc->file, &narc->num_files, 2);
         btnf_start = narc->btaf_start + chunk_size;
-        FS_SeekFile(&narc->file, btnf_start + 4, FS_SEEK_SET);
+        FS_SeekFile(&narc->file, (s32)(btnf_start + 4), FS_SEEK_SET);
         FS_ReadFile(&narc->file, &chunk_size, 4);
         narc->gmif_start = btnf_start + chunk_size;
     }
@@ -350,14 +350,14 @@ THUMB_FUNC void * NARC_AllocAndReadWholeMember(NARC * narc, u32 file_id, u32 hea
     void * dest;
     if (narc->num_files <= file_id)
         ErrorHandling();
-    FS_SeekFile(&narc->file, narc->btaf_start + 12 + 8 * file_id, FS_SEEK_SET);
+    FS_SeekFile(&narc->file, (s32)(narc->btaf_start + 12 + 8 * file_id), FS_SEEK_SET);
     FS_ReadFile(&narc->file, &file_start, 4);
     FS_ReadFile(&narc->file, &file_end, 4);
-    FS_SeekFile(&narc->file, narc->gmif_start + 8 + file_start, FS_SEEK_SET);
+    FS_SeekFile(&narc->file, (s32)(narc->gmif_start + 8 + file_start), FS_SEEK_SET);
     dest = AllocFromHeap(heap_id, file_end - file_start);
     if (dest != NULL)
     {
-        FS_ReadFile(&narc->file, dest, file_end - file_start);
+        FS_ReadFile(&narc->file, dest, (s32)(file_end - file_start));
     }
     return dest;
 }
@@ -368,11 +368,11 @@ THUMB_FUNC void NARC_ReadWholeMember(NARC * narc, u32 file_id, void * dest)
     u32 file_end;
     if (narc->num_files <= file_id)
         ErrorHandling();
-    FS_SeekFile(&narc->file, narc->btaf_start + 12 + 8 * file_id, FS_SEEK_SET);
+    FS_SeekFile(&narc->file, (s32)(narc->btaf_start + 12 + 8 * file_id), FS_SEEK_SET);
     FS_ReadFile(&narc->file, &file_start, 4);
     FS_ReadFile(&narc->file, &file_end, 4);
-    FS_SeekFile(&narc->file, narc->gmif_start + 8 + file_start, FS_SEEK_SET);
-    FS_ReadFile(&narc->file, dest, file_end - file_start);
+    FS_SeekFile(&narc->file, (s32)(narc->gmif_start + 8 + file_start), FS_SEEK_SET);
+    FS_ReadFile(&narc->file, dest, (s32)(file_end - file_start));
 }
 
 THUMB_FUNC u32 NARC_GetMemberSize(NARC * narc, u32 file_id)
@@ -381,7 +381,7 @@ THUMB_FUNC u32 NARC_GetMemberSize(NARC * narc, u32 file_id)
     u32 file_end;
     if (narc->num_files <= file_id)
         ErrorHandling();
-    FS_SeekFile(&narc->file, narc->btaf_start + 12 + 8 * file_id, FS_SEEK_SET);
+    FS_SeekFile(&narc->file, (s32)(narc->btaf_start + 12 + 8 * file_id), FS_SEEK_SET);
     FS_ReadFile(&narc->file, &file_start, 4);
     FS_ReadFile(&narc->file, &file_end, 4);
     return file_end - file_start;
@@ -392,15 +392,15 @@ THUMB_FUNC void NARC_ReadFromMember(NARC * narc, u32 file_id, u32 pos, u32 size,
     u32 file_start;
     if (narc->num_files <= file_id)
         ErrorHandling();
-    FS_SeekFile(&narc->file, narc->btaf_start + 12 + 8 * file_id, FS_SEEK_SET);
+    FS_SeekFile(&narc->file, (s32)(narc->btaf_start + 12 + 8 * file_id), FS_SEEK_SET);
     FS_ReadFile(&narc->file, &file_start, 4);
-    FS_SeekFile(&narc->file, narc->gmif_start + 8 + file_start + pos, FS_SEEK_SET);
-    FS_ReadFile(&narc->file, dest, size);
+    FS_SeekFile(&narc->file, (s32)(narc->gmif_start + 8 + file_start + pos), FS_SEEK_SET);
+    FS_ReadFile(&narc->file, dest, (s32)size);
 }
 
 THUMB_FUNC void NARC_ReadFile(NARC * narc, u32 size, void * dest)
 {
-    FS_ReadFile(&narc->file, dest, size);
+    FS_ReadFile(&narc->file, dest, (s32)size);
 }
 
 THUMB_FUNC u16 NARC_GetFileCount(NARC * narc)
